### 2023-05-30 #5502 @kthyng
Hi @giswqs and thanks for your submission. I saw your note that this package is newer than our guidelines suggest but you wanted to try submitting anyway — that's fine, but now I need to check with the greater editorial board to have a vote for it being in scope for JOSS. This will take 1-2 weeks. Thanks.
