### 2023-06-13 #5502 @kthyng
@giswqs The editorial board was positive about this going forward! I'll start looking for an editor.
