### 2023-06-14 b0c7085
Add rectangularize option for vector output
