### 2023-06-14 acd6a1d
Rename regularize function
