+++ samgeo/common.py
@@ -1717,7 +1717,7 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
 
     rectangular = widgets.Checkbox(
         value=False,
-        description="Rectangularize",
+        description="Regularize",
         layout=widgets.Layout(width="130px", padding=padding),
         style=style,
     )
@@ -1967,8 +1967,8 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
                         )
                         if m.find_layer("Masks") is not None:
                             m.remove_layer(m.find_layer("Masks"))
-                        if m.find_layer("Rectangular") is not None:
-                            m.remove_layer(m.find_layer("Rectangular"))
+                        if m.find_layer("Regularized") is not None:
+                            m.remove_layer(m.find_layer("Regularized"))
 
                         if hasattr(sam, "prediction_fp") and os.path.exists(
                             sam.prediction_fp
@@ -1988,14 +1988,16 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
 
                             if rectangular.value:
                                 vector = filename.replace(".tif", ".gpkg")
-                                vector_rec = filename.replace(".tif", "_rec.gpkg")
+                                vector_rec = filename.replace(".tif", "_rect.gpkg")
                                 raster_to_vector(filename, vector)
-                                rectangularize(vector, vector_rec)
+                                regularize(vector, vector_rec)
                                 vector_style = {"color": colorpicker.value}
                                 m.add_vector(
                                     vector_rec,
-                                    layer_name="Rectangular",
+                                    layer_name="Regularized",
                                     style=vector_style,
+                                    info_mode=None,
+                                    zoom_to_layer=False,
                                 )
 
                         except:
@@ -2018,8 +2020,8 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
                     vector = filename.replace(".tif", ".gpkg")
                     raster_to_vector(filename, vector)
                     if rectangular.value:
-                        vector_rec = filename.replace(".tif", "_rec.gpkg")
-                        rectangularize(vector, vector_rec)
+                        vector_rec = filename.replace(".tif", "_rect.gpkg")
+                        regularize(vector, vector_rec)
 
                     fg_points = [
                         [marker.location[1], marker.location[0]]
@@ -2080,8 +2082,8 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
             output.clear_output()
             try:
                 m.remove_layer(m.find_layer("Masks"))
-                if m.find_layer("Rectangular") is not None:
-                    m.remove_layer(m.find_layer("Rectangular"))
+                if m.find_layer("Regularized") is not None:
+                    m.remove_layer(m.find_layer("Regularized"))
                 m.clear_drawings()
                 if hasattr(m, "fg_markers"):
                     m.user_rois = None
@@ -2323,7 +2325,7 @@ def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
         description="Box threshold:",
         min=0,
         max=1,
-        value=0.5,
+        value=0.25,
         step=0.01,
         readout=True,
         continuous_update=True,
@@ -2336,7 +2338,7 @@ def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
         min=0,
         max=1,
         step=0.01,
-        value=0.5,
+        value=0.25,
         readout=True,
         continuous_update=True,
[…]
