### 2023-06-14 38c7346
Added regularized option for vector output (#104)
* Add rectangularize option for vector output

* Rename regularize function
