+++ samgeo/common.py
@@ -1714,11 +1714,28 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
         layout=widgets.Layout(width=widget_width, padding=padding),
         style=style,
     )
+
+    rectangular = widgets.Checkbox(
+        value=False,
+        description="Regularize",
+        layout=widgets.Layout(width="130px", padding=padding),
+        style=style,
+    )
+
+    colorpicker = widgets.ColorPicker(
+        concise=False,
+        description="Color",
+        value="#ffff00",
+        layout=widgets.Layout(width="140px", padding=padding),
+        style=style,
+    )
+
     buttons = widgets.VBox(
         [
             radio_buttons,
             widgets.HBox([fg_count, bg_count]),
             opacity_slider,
+            widgets.HBox([rectangular, colorpicker]),
             widgets.HBox(
                 [segment_button, save_button, reset_button],
                 layout=widgets.Layout(padding="0px 4px 0px 4px"),
@@ -1950,6 +1967,8 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
                         )
                         if m.find_layer("Masks") is not None:
                             m.remove_layer(m.find_layer("Masks"))
+                        if m.find_layer("Regularized") is not None:
+                            m.remove_layer(m.find_layer("Regularized"))
 
                         if hasattr(sam, "prediction_fp") and os.path.exists(
                             sam.prediction_fp
@@ -1966,6 +1985,21 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
                                 layer_name="Masks",
                                 zoom_to_layer=False,
                             )
+
+                            if rectangular.value:
+                                vector = filename.replace(".tif", ".gpkg")
+                                vector_rec = filename.replace(".tif", "_rect.gpkg")
+                                raster_to_vector(filename, vector)
+                                regularize(vector, vector_rec)
+                                vector_style = {"color": colorpicker.value}
+                                m.add_vector(
+                                    vector_rec,
+                                    layer_name="Regularized",
+                                    style=vector_style,
+                                    info_mode=None,
+                                    zoom_to_layer=False,
+                                )
+
                         except:
                             pass
                         output.clear_output()
@@ -1984,7 +2018,10 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
                     filename = chooser.selected
                     shutil.copy(sam.prediction_fp, filename)
                     vector = filename.replace(".tif", ".gpkg")
-                    raster_to_gpkg(filename, vector)
+                    raster_to_vector(filename, vector)
+                    if rectangular.value:
+                        vector_rec = filename.replace(".tif", "_rect.gpkg")
+                        regularize(vector, vector_rec)
 
                     fg_points = [
                         [marker.location[1], marker.location[0]]
@@ -2040,9 +2077,13 @@ def sam_map_gui(sam, basemap="SATELLITE", repeat_mode=True, out_dir=None, **kwar
             segment_button.value = False
             reset_button.value = False
             opacity_slider.value = 0.5
+            rectangular.value = False
+            colorpicker.value = "#ffff00"
             output.clear_output()
             try:
                 m.remove_layer(m.find_layer("Masks"))
+                if m.find_layer("Regularized") is not None:
+                    m.remove_layer(m.find_layer("Regularized"))
                 m.clear_drawings()
[…]
