### 2023-06-15 c5811e8
Add swimming pools example
