### 2023-06-15 2a52147
Added swimming pools example (#106)
