### 2023-06-16 7564ac3
Used sorted function instead of if statements (#109)
