### 2023-06-16 41d3692
Removed 'image' argument - was not being used in the show_box function (#111)
