### 2023-06-21 0983bbc
Add support for batch segmentation
