+++ samgeo/common.py
@@ -2257,7 +2257,16 @@ def install_package(package):
         process.wait()
 
 
-def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
+def text_sam_gui(
+    sam,
+    basemap="SATELLITE",
+    out_dir=None,
+    box_threshold=0.25,
+    text_threshold=0.25,
+    cmap="viridis",
+    opacity=0.5,
+    **kwargs,
+):
     """Display the SAM Map GUI.
 
     Args:
@@ -2323,7 +2332,7 @@ def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
         description="Box threshold:",
         min=0,
         max=1,
-        value=0.25,
+        value=box_threshold,
         step=0.01,
         readout=True,
         continuous_update=True,
@@ -2336,7 +2345,7 @@ def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
         min=0,
         max=1,
         step=0.01,
-        value=0.25,
+        value=text_threshold,
         readout=True,
         continuous_update=True,
         layout=widgets.Layout(width=widget_width, padding=padding),
@@ -2346,7 +2355,7 @@ def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
     cmap_dropdown = widgets.Dropdown(
         description="Palette:",
         options=cm.list_colormaps(),
-        value="viridis",
+        value=cmap,
         style=style,
         layout=widgets.Layout(width=widget_width, padding=padding),
     )
@@ -2355,7 +2364,7 @@ def text_sam_gui(sam, basemap="SATELLITE", out_dir=None, **kwargs):
         description="Opacity:",
         min=0,
         max=1,
-        value=0.5,
+        value=opacity,
         readout=True,
         continuous_update=True,
         layout=widgets.Layout(width=widget_width, padding=padding),
@@ -2627,3 +2636,160 @@ def regularize(source, output=None, crs="EPSG:4326", **kwargs):
         result.to_file(output, **kwargs)
     else:
         return result
+
+
+def split_raster(filename, out_dir, tile_size=256, overlap=0):
+    """Split a raster into tiles.
+
+    Args:
+        filename (str): The path or http URL to the raster file.
+        out_dir (str): The path to the output directory.
+        tile_size (int | tuple, optional): The size of the tiles. Can be an integer or a tuple of (width, height). Defaults to 256.
+        overlap (int, optional): The number of pixels to overlap between tiles. Defaults to 0.
+
+    Raises:
+        ImportError: Raised if GDAL is not installed.
+    """
+
+    try:
+        from osgeo import gdal
+    except ImportError:
+        raise ImportError(
+            "GDAL is required to use this function. Install it with `conda install gdal -c conda-forge`"
+        )
+
+    if isinstance(filename, str):
+        if filename.startswith("http"):
+            output = filename.split("/")[-1]
+            download_file(filename, output)
+            filename = output
+
+    # Open the input GeoTIFF file
+    ds = gdal.Open(filename)
+
+    if not os.path.exists(out_dir):
+        os.makedirs(out_dir)
+
+    if isinstance(tile_size, int):
+        tile_width = tile_size
+        tile_height = tile_size
+    elif isinstance(tile_size, tuple):
+        tile_width = tile_size[0]
+        tile_height = tile_size[1]
+
+    # Get the size of the input raster
+    width = ds.RasterXSize
+    height = ds.RasterYSize
+
+    # Calculate the number of tiles needed in both directions, taking into account the overlap
+    num_tiles_x = (width - overlap) // (tile_width - overlap) + int(
[…]
