### 2023-06-21 0238f24
Added support for batch segmentation (#116)
