### 2023-06-21 b82f2ae
Fix mask bug in text_sam/predict when the input is PIL image (#117)
