### 2023-06-23 d465366
Fix file deletion error on Windows
