### 2023-06-23 7f0592d
Add text_sam to docs
