+++ docs/text_sam.md
@@ -1,0 +1,3 @@
+# text_sam module
+
+::: samgeo.text_sam
