### 2023-06-23 15a8211
Added text_sam to docs (#123)
* Add text_sam to docs

* Add text_sam to toc
