### 2023-06-28 b31acc6
Update text_sam.py
Added to return pixel coordinates of masks
