### 2023-06-28 c4e4bb3
Improve exception handling for LangSAM GUI
