### 2023-06-28 a962fc7
Fixed image source bug for show_anns (#131)
