### 2023-07-08 d083038
Fixed save_masks bug without source image (#139)
