### 2023-07-10 a6d13ac
Added info about libgl1 dependency install on Linux systems (#141)
