+++ changelog/changelog.md
@@ -1,0 +1,184 @@
+# Changelog
+
+## v0.8.4 - Jul 5, 2023
+
+**Improvements**
+
+-   Fixed model download bug (#136)
+-   Added legal notice (#133)
+-   Fixed image source bug for show_anns (#131)
+-   Improved exception handling for LangSAM GUI (#130)
+-   Added to return pixel coordinates of masks (#129)
+-   Added text_sam to docs (#123)
+-   Fixed file deletion error on Windows (#122)
+-   Fixed mask bug in text_sam/predict when the input is PIL image (#117)
+
+## v0.8.3 - Jun 20, 2023
+
+**New Features**
+
+-   Added support for batch segmentation (#116)
+-   Added swimming pools example (#106)
+
+**Improvements**
+
+-   Removed 'flag' and 'param' arguments (#112)
+-   Used sorted function instead of if statements (#109)
+
+## v0.8.2 - Jun 14, 2023
+
+**New Features**
+
+-   Added regularized option for vector output (#104)
+-   Added text prompt GUI (#80)
+
+**Improvements**
+
+-   Added more deep learning resources (#90)
+-   Use the force_filename parameter with hf_hub_download() (#93)
+-   Fixed typo (#94)
+
+## v0.8.1 - May 24, 2023
+
+**Improvements**
+
+-   Added huggingface_hub and remove onnx (#87)
+-   Added more demos to docs (#82)
+
+## v0.8.0 - May 24, 2023
+
+**New Features**
+
+-   Added support for using text prompts with SAM (#73)
+-   Added text prompt GUI (#80)
+
+**Improvements**
+
+-   Improved text prompt notebook (#79)
+-   Fixed notebook typos (#78)
+-   Added ArcGIS tutorial to docs (#72)
+
+## v0.7.0 - May 20, 2023
+
+**New Features**
+
+-   Added unittest (#58)
+-   Added JOSS paper draft (#61)
+-   Added ArcGIS notebook example (#63)
+-   Added text prompting segmentation (#65)
+-   Added support for segmenting non-georeferenced imagery (#66)
+
+**Improvements**
+
+-   Added blend option for show_anns method (#59)
+-   Updated ArcGIS installation instructions (#68, #70)
+
+**Contributors**
+
+@p-vdp @LucasOsco
+
+## v0.6.2 - May 17, 2023
+
+**Improvements**
+
+-   Added jupyter-server-proxy to Dockerfile for supporting add_raster (#57)
+
+## v0.6.1 - May 16, 2023
+
+**New Features**
+
+-   Added Dockerfile (#51)
+
+## v0.6.0 - May 16, 2023
+
+**New Features**
+
+-   Added interactive GUI for creating foreground and background markers (#44)
+-   Added support for custom projection bbox (#39)
+
+**Improvements**
+
+-   Fixed Colab Marker AwesomeIcon bug (#50)
+-   Added info about using SAM with Desktop GIS (#48)
+-   Use proper extension in the usage documentation (#43)
+
+**Demos**
+
+-   Interactive segmentation with input prompts
+
+![](https://i.imgur.com/2Nyg9uW.gif)
+
+-   Input prompts from existing files
+
+![](https://i.imgur.com/Cb4ZaKY.gif)
+
+## v0.5.0 - May 10, 2023
+
+**New Features**
+
+-   Added support for input prompts (#30)
+
+**Improvements**
+
+-   Fixed the batch processing bug (#29)
+
+**Demos**
+
+![](https://i.imgur.com/GV7Rzxt.gif)
+
+## v0.4.0 - May 6, 2023
+
+**New Features**
+
[…]
