+++ samgeo/samgeo.md
@@ -1,0 +1,4 @@
+ 
+# samgeo module
+
+::: samgeo.samgeo
