### 2023-07-19 #5663 @editorialbot
[REVIEW]: samgeo: A Python package for segmenting geospatial data with the Segment Anything Model (SAM)
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** https://github.com/opengeos/segment-geospatial
**Branch with paper.md** (empty if default branch):
**Version:** v0.10.1
