### 2023-07-19 c60f1c8
Updated installation docs (#146)
