### 2023-07-19 #5663 @ayoubft
> ## Reviewer instructions & questions
>
> @ayoubft & @aazuspan, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
> First of all you need to run this command in a separate comment to create the checklist:
>
> ```
> @editorialbot generate my checklist
> ```

@giswqs, actually the reviewer have not yet generated his checklist to start reviewing. Once he will comment the command above, the comment will be edited to the checklist that the reviewer will be crossing each bullet as done.
