### 2023-07-22 opengeos/segment-geospatial#150 @aazuspan
JOSS review
### Description

Hi @giswqs, I'm going through the JOSS review checklist for https://github.com/openjournals/joss-reviews/issues/5663 and had a few comments.

On the [PDF draft](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05663/joss.05663/10.21105.joss.05663.pdf):

- [x] Line 13: Extra or misplaced word "and" in the sentence "... to provide a straightforward interface for users to segment and remote sensing imagery ..."
- [x] Line 14: Just a suggestion, but you could consider mentioning in the summary that segmentation can be run automatically, interactively in a GUI, or by text prompt - those seem like major features. Of course if text prompting was mentioned, the Grounding DINO paper would need to be cited.

Other things:

- [x] The contributing guide references testing in some [old Python versions](https://github.com/opengeos/segment-geospatial/blob/3fe3d3459564946297cdbe51a7f75998116dc88f/docs/contributing.md?plain=1#L106) that don't match the CI workflows. The guide also references `README.rst`, which presumably should be `README.md` instead.

I'll let you know if I have any other comments, but that's all that popped out. Thanks!
