### 2023-07-22 a936069
Address JOSS reviewer comments
