+++ paper/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 # Summary
 
-Segment-Geospatial (samgeo) is an open-source Python package designed to simplify the process of segmenting geospatial data with the Segment Anything Model [@Kirillov2023]. The package leverages popular Python libraries, such as leafmap [@Wu2021], ipywidgets [@Grout2021], rasterio [@Gillies2013], geopandas [@Jordahl2021], and segment-anything-py [@Wu2023], to provide a straightforward interface for users to segment and remote sensing imagery and export the results in various formats, including vector and raster data. It also provides functionality for downloading remote sensing imagery and visualizing segmentation results interactively in a Jupyter environment. Segment-Geospatial aims to fill the gap in the Python ecosystem by providing a user-friendly, efficient, and flexible geospatial segmentation tool without the need for training deep learning models.
+Segment-Geospatial (samgeo) is an open-source Python package designed to simplify the process of segmenting geospatial data with the Segment Anything Model [@Kirillov2023]. The package leverages popular Python libraries, such as leafmap [@Wu2021], ipywidgets [@Grout2021], rasterio [@Gillies2013], geopandas [@Jordahl2021], and segment-anything-py [@Wu2023], to provide a straightforward interface for users to segment remote sensing imagery and export the results in various formats, including vector and raster data. The segmentation can be run automatically, interactively in a graphical user interface (GUI), or by text prompts built upon Grounding DINO [@liu2023]. It also provides functionality for downloading remote sensing imagery and visualizing segmentation results interactively in a Jupyter environment. Segment-Geospatial aims to fill the gap in the Python ecosystem by providing a user-friendly, efficient, and flexible geospatial segmentation tool without the need for training deep learning models.
 
 # Statement of need
 
