+++ paper/paper.bib
@@ -113,3 +113,13 @@
       doi={10.48550/arXiv.2304.05750},
       url={https://doi.org/10.48550/arXiv.2304.05750}
 }
+
+
+@ARTICLE{liu2023,
+      title={Grounding DINO: Marrying DINO with Grounded Pre-Training for Open-Set Object Detection}, 
+      author={Shilong Liu and Zhaoyang Zeng and Tianhe Ren and Feng Li and Hao Zhang and Jie Yang and Chunyuan Li and Jianwei Yang and Hang Su and Jun Zhu and Lei Zhang},
+      year={2023},
+      eprint={2303.05499},
+      archivePrefix={arXiv},
+      primaryClass={cs.CV}
+}
