### 2023-07-22 opengeos/segment-geospatial#150 @giswqs
@aazuspan Thank you for the feedback. This has been addressed in #151
