### 2023-07-22 opengeos/segment-geospatial#150 @aazuspan
Great, I'm satisfied this fits all the JOSS criteria. Thanks @giswqs, it's a very cool package!

p.s. It looks like you may need to regenerate the PDF on the review issue.

> For example, to regenerate the paper pdf after making changes in the paper's md or bib files, type:
