+++ paper/paper.md
@@ -39,4 +39,6 @@ Traditionally, image segmentation has been performed using manual or semi-automa
 
 The segment-geospatial package draws its inspiration from segment-anything-eo repository [@Hancharenka2023]. We thank the author Aliaksandr Hancharenka for his excellent work. We also thank the authors of the Segment Anything Model [@Kirillov2023] for making the model available to the public.
 
+This project is based upon work partially supported by the National Aeronautics and Space Administration (NASA) under Grant No. 80NSSC22K1742 issued through the [Open Source Tools, Frameworks, and Libraries 2020 Program](https://bit.ly/3RVBRcQ). This project is also supported by Amazon Web Services ([AWS](https://aws.amazon.com/)).
+
 # References
