### 2023-08-06 3753c2b
Add functions for saving bbox coords to vector
