### 2023-08-06 7ef4680
Add save_boxes method
