+++ samgeo/samgeo.py
@@ -528,6 +528,12 @@ class SamGeo:
 
         """
 
+        if isinstance(boxes, str):
+            gdf = gpd.read_file(boxes)
+            if gdf.crs is not None:
+                gdf = gdf.to_crs("epsg:4326")
+            boxes = gdf.geometry.bounds.values.tolist()
+
         if isinstance(point_coords, str):
             point_coords = vector_to_geojson(point_coords)
 
