### 2023-08-06 a16753f
Add geojson support for boxes
