### 2023-08-06 1d7b42a
Added support for box prompts (#159)
* Add support for multiple input boxes

* Clean up code

* Add functions for saving bbox coords to vector

* Add save_boxes method

* Add geojson support for boxes

* Add notebook example

* Update README

* Prep for v0.9.0
