+++ samgeo/samgeo.py
@@ -486,7 +486,7 @@ class SamGeo:
         self,
         point_coords=None,
         point_labels=None,
-        box=None,
+        boxes=None,
         point_crs=None,
         mask_input=None,
         multimask_output=True,
@@ -494,7 +494,7 @@ class SamGeo:
         output=None,
         index=None,
         mask_multiplier=255,
-        dtype=np.float32,
+        dtype="float32",
         return_results=False,
         **kwargs,
     ):
@@ -507,7 +507,7 @@ class SamGeo:
             point_labels (list | int | np.ndarray, optional): A length N array of labels for the
                 point prompts. 1 indicates a foreground point and 0 indicates a background point.
             point_crs (str, optional): The coordinate reference system (CRS) of the point prompts.
-            box (list | np.ndarray, optional): A length 4 array given a box prompt to the
+            boxes (list | np.ndarray, optional): A length 4 array given a box prompt to the
                 model, in XYXY format.
             mask_input (np.ndarray, optional): A low resolution mask input to the model, typically
                 coming from a previous prediction iteration. Has form 1xHxW, where for SAM, H=W=256.
@@ -528,6 +528,18 @@ class SamGeo:
 
         """
 
+        if isinstance(boxes, str):
+            gdf = gpd.read_file(boxes)
+            if gdf.crs is not None:
+                gdf = gdf.to_crs("epsg:4326")
+            boxes = gdf.geometry.bounds.values.tolist()
+        elif isinstance(boxes, dict):
+            import json
+
+            geojson = json.dumps(boxes)
+            gdf = gpd.read_file(geojson, driver="GeoJSON")
+            boxes = gdf.geometry.bounds.values.tolist()
+
         if isinstance(point_coords, str):
             point_coords = vector_to_geojson(point_coords)
 
@@ -540,16 +552,17 @@ class SamGeo:
         if hasattr(self, "point_labels"):
             point_labels = self.point_labels
 
-        if point_crs is not None:
+        if (point_crs is not None) and (point_coords is not None):
             point_coords = coords_to_xy(self.source, point_coords, point_crs)
 
         if isinstance(point_coords, list):
             point_coords = np.array(point_coords)
 
-        if point_labels is None:
-            point_labels = [1] * len(point_coords)
-        elif isinstance(point_labels, int):
-            point_labels = [point_labels] * len(point_coords)
+        if point_coords is not None:
+            if point_labels is None:
+                point_labels = [1] * len(point_coords)
+            elif isinstance(point_labels, int):
+                point_labels = [point_labels] * len(point_coords)
 
         if isinstance(point_labels, list):
             if len(point_labels) != len(point_coords):
@@ -561,22 +574,115 @@ class SamGeo:
                     )
             point_labels = np.array(point_labels)
 
-        if isinstance(box, list) and point_crs is not None:
-            box = np.array(bbox_to_xy(self.source, box, point_crs))
-
         predictor = self.predictor
-        masks, scores, logits = predictor.predict(
-            point_coords, point_labels, box, mask_input, multimask_output, return_logits
-        )
+
+        input_boxes = None
+        if isinstance(boxes, list) and (point_crs is not None):
+            coords = bbox_to_xy(self.source, boxes, point_crs)
+            input_boxes = np.array(coords)
+            if isinstance(coords[0], int):
+                input_boxes = input_boxes[None, :]
+            else:
+                input_boxes = torch.tensor(input_boxes, device=self.device)
+                input_boxes = predictor.transform.apply_boxes_torch(
+                    input_boxes, self.image.shape[:2]
+                )
+        elif isinstance(boxes, list) and (point_crs is None):
+            input_boxes = np.array(boxes)
+            if isinstance(boxes[0], int):
+                input_boxes = input_boxes[None, :]
+
+        self.boxes = input_boxes
+
[…]
