+++ mkdocs.yml
@@ -52,6 +52,7 @@ nav:
           - examples/satellite.ipynb
           - examples/automatic_mask_generator.ipynb
           - examples/input_prompts.ipynb
+          - examples/box_prompts.ipynb
           - examples/text_prompts.ipynb
           - examples/text_prompts_batch.ipynb
           - examples/swimming_pools.ipynb
