+++ samgeo/__init__.py
@@ -1,8 +1,8 @@
 """Top-level package for segment-geospatial."""
 
 __author__ = """Qiusheng Wu"""
-__email__ = 'giswqs@gmail.com'
-__version__ = '0.8.5'
+__email__ = "giswqs@gmail.com"
+__version__ = "0.8.5"
 
 
 from .samgeo import *
