### 2023-08-06 e018ff1
Fixed bumpversion error (#160)
* Update notebook

* Update init.py

* Update setup.py

* Update init.py
