+++ setup.py
@@ -57,6 +57,6 @@ setup(
     test_suite="tests",
     tests_require=test_requirements,
     url="https://github.com/opengeos/segment-geospatial",
-    version="0.8.5",
+    version='0.8.5',
     zip_safe=False,
 )
