+++ samgeo/__init__.py
@@ -1,7 +1,7 @@
 """Top-level package for segment-geospatial."""
 
 __author__ = """Qiusheng Wu"""
-__email__ = 'giswqs@gmail.com'
+__email__ = "giswqs@gmail.com"
 __version__ = '0.8.5'
 
 
