+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.8.5
+current_version = 0.9.0
 commit = True
 tag = True
 
