### 2023-08-13 3000c8e
Revise download_checkpoint function
