### 2023-08-13 2439f59
Add hq_sam module
