+++ requirements.txt
@@ -11,3 +11,4 @@ xyzservices
 pyproj
 leafmap
 localtileserver
+timm
