+++ docs/hq_sam.md
@@ -1,0 +1,3 @@
+# hq_sam module
+
+::: samgeo.hq_sam
