### 2023-08-13 ef1c4dd
Added support for HQ-SAM (#161)
* Revise download_checkpoint function

* Add hq_sam module
