+++ mkdocs.yml
@@ -60,4 +60,5 @@ nav:
     - API Reference:
           - common module: common.md
           - samgeo module: samgeo.md
+          - hq_sam module: hq_sam.md
           - text_sam module: text_sam.md
