### 2023-08-19 dd691e3
Improved contributing guidelines (#169)
