### 2023-08-22 95c6c09
[FIX] Added missing conversion from BGR to RGB after reading an image using cv2.imread
