### 2023-08-22 opengeos/segment-geospatial#174 @ayoubft
JOSS Review - @ayoubft
Hi @giswqs! I'm currently working my way through the JOSS review checklist for https://github.com/openjournals/joss-reviews/issues/5663, and I've got a few things to discuss.

To start, I want to express my heartfelt thanks for all the hard work you've put into developing these geospatial tools. Your efforts are truly valued and appreciated.

The points:
- [x]  On line 38, you mention that there are no Python packages for geospatial imagery segmentation, which is technically correct. However, it might be a bit misleading because there is some inspiration drawn from aliaksandr960's work, as you mentioned on line 42.
- [x]  I've given the main functionalities a try based on your tutorials, and I noticed that while some prompts work well, others might need a bit of parameter fine-tuning. It might be a good idea to mention this in the summary, especially for text prompting.
- [x] A friendly suggestion: Consider using GitHub CI to automate your testing processes. It could streamline your workflow!

I'll reach out if any other thoughts come to mind, but for now, that's all I wanted to mention. Thanks!
