### 2023-08-22 d77bc91
Address JOSS review comments (#175)
