### 2023-08-22 opengeos/segment-geospatial#174 @giswqs
@ayoubft Thank you for the feedback. Points 1&2 have been addressed in #175.

GitHub CI has already been utilize to streamline the testing processes. Keep in mind that some functions can take a long time to execute without a GPU. Therefore, not all tests are feasible with GitHub CI.
https://github.com/opengeos/segment-geospatial/blob/main/.github/workflows/ubuntu.yml#L50
