+++ samgeo/fast_sam.py
@@ -2,6 +2,8 @@
 https://github.com/opengeos/FastSAM
 """
 
+import matplotlib.pyplot as plt
+import numpy as np
 import torch
 from .common import *
 
@@ -45,12 +47,13 @@ class SamGeo(FastSAM):
 
         super().__init__(model, **kwargs)
 
-    def set_image(self, image, device=None):
+    def set_image(self, image, device=None, **kwargs):
         """Set the input image.
 
         Args:
             image (str): The path to the image file or a HTTP URL.
             device (str, optional): The device to use. Defaults to "cuda" if available, otherwise "cpu".
+            kwargs: Additional keyword arguments to pass to the FastSAM model.
         """
 
         if isinstance(image, str):
@@ -70,11 +73,215 @@ class SamGeo(FastSAM):
             if device == "cuda":
                 torch.cuda.empty_cache()
 
-        everything_results = self(image)
+        everything_results = self(image, device=device, **kwargs)
 
         self.prompt_process = FastSAMPrompt(image, everything_results, device=device)
 
-    def everything_prompt(self, output=None):
+    def everything_prompt(self, output=None, **kwargs):
+        """Segment the image with the everything prompt. Adapted from
+        https://github.com/CASIA-IVA-Lab/FastSAM/blob/main/fastsam/prompt.py#L451
+
+        Args:
+            output (str, optional): The path to save the output image. Defaults to None.
+        """
+
         prompt_process = self.prompt_process
         ann = prompt_process.everything_prompt()
-        prompt_process.plot(annotations=ann, output_path=output)
+        self.annotations = ann
+
+        if output is not None:
+            self.save_masks(output, **kwargs)
+        else:
+            return ann
+
+    def point_prompt(self, points, pointlabel, output=None, **kwargs):
+        """Segment the image with the point prompt. Adapted from
+        https://github.com/CASIA-IVA-Lab/FastSAM/blob/main/fastsam/prompt.py#L414
+
+        Args:
+            points (list): A list of points.
+            pointlabel (list): A list of labels for each point.
+            output (str, optional): The path to save the output image. Defaults to None.
+        """
+
+        prompt_process = self.prompt_process
+        ann = prompt_process.point_prompt(points, pointlabel)
+        self.annotations = ann
+
+        if output is not None:
+            self.save_masks(output, **kwargs)
+        else:
+            return ann
+
+    def box_prompt(self, bbox=None, bboxes=None, output=None, **kwargs):
+        """Segment the image with the box prompt. Adapted from
+        https://github.com/CASIA-IVA-Lab/FastSAM/blob/main/fastsam/prompt.py#L377
+
+        Args:
+            bbox (list, optional): The bounding box. Defaults to None.
+            bboxes (list, optional): A list of bounding boxes. Defaults to None.
+            output (str, optional): The path to save the output image. Defaults to None.
+        """
+
+        prompt_process = self.prompt_process
+        ann = prompt_process.box_prompt(bbox, bboxes)
+        self.annotations = ann
+
+        if output is not None:
+            self.save_masks(output, **kwargs)
+        else:
+            return ann
+
+    def text_prompt(self, text, output=None, **kwargs):
+        """Segment the image with the text prompt. Adapted from
+        https://github.com/CASIA-IVA-Lab/FastSAM/blob/main/fastsam/prompt.py#L439
+
+        Args:
+            text (str): The text to segment.
+            output (str, optional): The path to save the output image. Defaults to None.
+        """
+
+        prompt_process = self.prompt_process
+        ann = prompt_process.text_prompt(text)
+        self.annotations = ann
+
+        if output is not None:
[…]
