+++ mkdocs.yml
@@ -62,5 +62,6 @@ nav:
     - API Reference:
           - common module: common.md
           - samgeo module: samgeo.md
+          - fast_sam module: fast_sam.md
           - hq_sam module: hq_sam.md
           - text_sam module: text_sam.md
