### 2023-08-24 c938d37
Add show_anns function
