### 2023-08-24 c0956ad
Added fastsam module (#167)
* Add fastsam module

* Add fastsam functions and notebook

* Add show_anns function

* Update dep
