### 2023-08-24 f72da1e
Add fastsam notebook to docs
