+++ samgeo/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Qiusheng Wu"""
 __email__ = "giswqs@gmail.com"
-__version__ = '0.9.1'
+__version__ = '0.10.0'
 
 
 from .samgeo import *
