+++ docs/fast_sam.md
@@ -1,0 +1,3 @@
+# fast_sam module
+
+::: samgeo.fast_sam
