+++ mkdocs.yml
@@ -57,6 +57,7 @@ nav:
           - examples/box_prompts.ipynb
           - examples/text_prompts.ipynb
           - examples/text_prompts_batch.ipynb
+          - examples/fast_sam.ipynb
           - examples/swimming_pools.ipynb
           - examples/arcgis.ipynb
     - API Reference:
