### 2023-08-27 opengeos/segment-geospatial#179 @SimonMolinsky
JOSS Review - Final remarks
Hi @giswqs ,

I've checked everything, and the package is outstanding! There are minor glitches that can be addressed to make using the package more accessible for newcomers:

1. I usually look into `README` to check how to install the package. Could you add a section below `Demos` or a link to the installation procedure in Readme?
2. In the installation example with conda, you use `Python=3.10`. I was wondering if it is the minimal version required, so I installed and tested the package with `Python=3.8`, and it worked fine. Is there a reason to fix the version of Python in the installation example?

Code quality is important to me, and I enjoyed how clean and concise your code is. You have provided excellent examples, and API is easy to grasp. Thank you for your work.
