### 2023-08-27 #5663 @SimonMolinsky
Hi @giswqs ,

I've created the new (minor) issue in your repository: https://github.com/opengeos/segment-geospatial/issues/179, but it doesn't affect the review. **My review is done**. Thank you for the excellent tool @giswqs
