### 2023-08-27 88ed304
Add installation instructions to README
