### 2023-08-27 opengeos/segment-geospatial#179 @giswqs
@SimonMolinsky Thank you for your feedback. The issues have been fixed in #180

- An installation section have been added to README
- The package supports Python >=3.8, which is specificed [here](https://github.com/opengeos/segment-geospatial/blob/main/setup.py#L34). I have removed the version pin 3.10 in the installation instructions. Without the pin, conda will usually install the latest version. For now, it is 3.11. Since 3.11 has been out for a while, all samgeo package dependencies have now supported Python 3.11, so it is safe to remove the 3.10 version pin.
