### 2023-08-27 #5663 @giswqs
@SimonMolinsky Thank you again for your feedback. The issues have been addressed in https://github.com/opengeos/segment-geospatial/pull/180
