### 2023-08-30 738f235
Fixed single bbox bug (#184)
