### 2023-09-12 #5663 @hugoledoux
oh wow, I went on holidays 3 weeks and didn't expect that all would be finish!? Thanks to all 3 reviewers!

I just read the paper again and cannot find any typos/mistakes. I wish more submissions to JOSS were like yours @giswqs ❤️

At this point could you:
