### 2023-09-12 #5663 @giswqs
@hugoledoux I have completed the tasks as requested. Check the links below. Thank you.

Tag release: [v0.10.1](https://github.com/opengeos/segment-geospatial/releases/tag/v0.10.1)
DOI: [10.5281/zenodo.8338074](https://doi.org/10.5281/zenodo.8338074)
