### 2023-09-12 #5663 @kthyng
Congrats on your new publication @giswqs! Many thanks to editor @hugoledoux and reviewers @ayoubft, @aazuspan, and @SimonMolinsky for your time, hard work, and expertise!!
